requirements.txt
pytest==5.4.1
pytest-html==2.1.1
requests==2.23.0
attrs==21.2.0